# backend/openapi.yaml
openapi: 3.0.3
info:
  title: NutriFlow API
  version: 1.0.0
  description: |
    Официальная спецификация NutriFlow API, синхронизированная с текущей реализацией.
    Покрывает аутентификацию, клиентов, дэшборд, меню, события, анализы и биллинг.
servers:
  - url: http://localhost:4000/api
    description: Local dev

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }

    ClientSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        goal: { type: string, nullable: true }
        status:
          type: string
          enum: [active, paused]
        proteinCoverage: { type: number }
        fiberCoverage: { type: number }
        kcalCoverage: { type: number }
        riskFlags:
          type: array
          items:
            type: string

    ClientNorms:
      type: object
      properties:
        kcalMin: { type: integer, nullable: true }
        kcalMax: { type: integer, nullable: true }
        proteinGrams: { type: integer, nullable: true }
        fatGramsMin: { type: integer, nullable: true }
        fatGramsMax: { type: integer, nullable: true }
        carbsGramsMin: { type: integer, nullable: true }
        carbsGramsMax: { type: integer, nullable: true }
        fiberGrams: { type: integer, nullable: true }

    ClientDayStats:
      type: object
      properties:
        id: { type: string }
        date: { type: string, format: date-time }
        kcal: { type: integer, nullable: true }
        protein: { type: integer, nullable: true }
        fat: { type: integer, nullable: true }
        carbs: { type: integer, nullable: true }
        fiber: { type: integer, nullable: true }
        kcalCoverage: { type: number, nullable: true }
        proteinCoverage: { type: number, nullable: true }
        fiberCoverage: { type: number, nullable: true }
        riskFlags:
          type: array
          items: { type: string }

    LabTest:
      type: object
      properties:
        id: { type: string }
        takenAt: { type: string, format: date-time }
        type: { type: string }
        marker: { type: string }
        value: { type: number }
        unit: { type: string }
        status:
          type: string
          enum: [LOW, NORMAL, HIGH]

    MenuTemplate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        focus: { type: string, nullable: true }

    MenuAssignment:
      type: object
      properties:
        id: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time, nullable: true }
        isActive: { type: boolean }
        menuTemplate:
          $ref: '#/components/schemas/MenuTemplate'

    Event:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string, nullable: true }
        title: { type: string }
        description: { type: string, nullable: true }
        type: { type: string }
        scheduledAt: { type: string, format: date-time }
        channel: { type: string, nullable: true }

    BillingPlan:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        maxClients: { type: integer }
        features:
          type: array
          items: { type: string }

    ClientProfile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status:
          type: string
          enum: [active, paused]
        goal: { type: string, nullable: true }
        norms:
          allOf:
            - $ref: '#/components/schemas/ClientNorms'
            - type: object
              nullable: true
        dayStats:
          allOf:
            - $ref: '#/components/schemas/ClientDayStats'
            - type: object
              nullable: true
        labs:
          type: array
          items:
            $ref: '#/components/schemas/LabTest'
        activeMenu:
          allOf:
            - $ref: '#/components/schemas/MenuAssignment'
            - type: object
              nullable: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    DashboardSummary:
      type: object
      properties:
        totalClients: { type: integer }
        activeClients: { type: integer }
        risks:
          type: object
          additionalProperties: { type: integer }
        eventsUpcomingCount: { type: integer }
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        menu:
          type: object
          properties:
            clientsWithActiveMenu: { type: integer }
        labs:
          type: object
          properties:
            lowRiskClients: { type: integer }

paths:
  /auth/login:
    post:
      summary: Login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '401':
          description: Invalid credentials

  /auth/me:
    get:
      summary: Current user
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /clients:
    get:
      summary: List clients with summary metrics
      tags: [clients]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientSummary'

  /clients/{id}/profile:
    get:
      summary: Aggregated client profile
      tags: [clients]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Client profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '404':
          description: Not found

  /clients/{id}/norms:
    put:
      summary: Update client nutrient norms
      tags: [clients]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientNorms'
      responses:
        '200':
          description: Updated client profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '400':
          description: Validation error
        '404':
          description: Client not found

  /dashboard/summary:
    get:
      summary: Dashboard summary
      tags: [dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'

  /clients/{id}/labs:
    get:
      summary: List lab tests for client
      tags: [labs]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Lab tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabTest'
    post:
      summary: Create lab test
      tags: [labs]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabTest'
      responses:
        '201':
          description: Created
        '400':
          description: Validation error

  /menu-templates:
    get:
      summary: List menu templates
      tags: [menu]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuTemplate'

  /clients/{id}/menu:
    get:
      summary: Get client menu assignments
      tags: [menu]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Menu assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    $ref: '#/components/schemas/MenuAssignment'
                  archived:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuAssignment'

  /clients/{id}/menu-assignment:
    post:
      summary: Assign menu template to client
      tags: [menu]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [menuTemplateId]
              properties:
                menuTemplateId: { type: string }
                startDate: { type: string, format: date-time, nullable: true }
                endDate: { type: string, format: date-time, nullable: true }
      responses:
        '200':
          description: Updated client profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'

  /clients/{id}/events:
    get:
      summary: List client events
      tags: [events]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Events
          content:
            application/json